services/user-service/requirements.txt
# Core dependencies for Lambda functions
# Database (same version as product-service)
psycopg2-binary==2.8.6
# Authentication (required for user-service)
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
